Add Delete Testing around new Viewsets
